add level transition
Includes a quick workaround for the issue where spamming step
causes the position to advance every time, ahead of the animation
and ahead of the win check which is triggered on animation finish.
The workaround is to not allow pressing the button while
animating, definitely not the best solution but the easiest to
deal with, and probably not all that bad.

Also introduces win delay, which takes care of a few issues of
what happens when the level to draw changes while drawing it.